requirement.txt
django
djangorestframework
environs
djangorestframework-simplejwt
python-decouple
pip install drf-yasg
django-extensions